requirements.txt
pyside6
simpleaudio
pillow
pyaudio
pyinstaller
